Reverted to trusting the data source to update the badge number
The first approach ended up being the correct one, because when I performed the change, I didn’t take into account that I also had to update the badge number when the user removes items from the list for example, or from the search. In any way, the data source knows exactly how many items is displaying, so it makes sense to trust it to update the badge.